+++ src/app/components/altmetric/altmetric.component.html
@@ -1,5 +1,4 @@
 <div class="layout-row">
-  {{ doi() }} {{ isbn() }}
   <ng-container *ngIf="shouldRender()">
     <div
       *ngIf="doi()"
@@ -9,7 +8,7 @@
       data-link-target="_blank"
       rel="noopener"
       class="altmetric-embed"
-      data-hide-no-mentions="false"
+      data-hide-no-mentions="true"
       data-badge-type="4"
       data-badge-popover="right"
       [attr.data-doi]="doi()"></div>
@@ -21,7 +20,7 @@
       data-link-target="_blank"
       rel="noopener"
       class="altmetric-embed"
-      data-hide-no-mentions="false"
+      data-hide-no-mentions="true"
       data-badge-type="4"
       data-badge-popover="right"
       [attr.data-isbn]="isbn()"></div>
