+++ src/app/components/altmetric/altmetric.component.scss
@@ -1,3 +1,12 @@
 :host {
-  display: block;
+  display: flex;
+  align-items: center;
+  height: 100%;
+}
+
+.layout-row {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 0.5rem;
 }
